refactor: Use native Bevy IVec2 for all geometry manipulations.
This allows easy geometric calculations which are very usual in Bevy.
Additionally the window coordinate system is now absolute between the
displays.

+++ src/config.rs
@@ -487,8 +487,8 @@ pub struct WindowParams {
     pub floating: Option<bool>,
     /// An optional preferred index for the window's position in the window strip.
     pub index: Option<usize>,
-    pub vertical_padding: Option<u16>,
-    pub horizontal_padding: Option<u16>,
+    pub vertical_padding: Option<i32>,
+    pub horizontal_padding: Option<i32>,
     pub dont_focus: Option<bool>,
 }
